pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cm360-bigquery"
version = "0.1.0"
description = "Reusable Gmail attachment to BigQuery ingestion service"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.9",
  "PyYAML>=6.0.0",
  "google-api-python-client>=2.154.0",
  "google-auth>=2.36.0",
  "google-auth-httplib2>=0.2.0",
  "google-cloud-bigquery>=3.28.0",
]

[project.scripts]
ingest-run = "src.main:run_cli"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]

